Fixed QR code crash bug
Fixed the crash bug. The crash occured when the texture stream stopped streaming. (webcam passtrough texture). This was because of the unity version we were using (2020.3.30f1). This unity version had a bug with this matter. That is why we changed the unity version to 2019.4.36f1. This version does not have this issue and supports everything we need for the project.